requirements.txt
mem0ai>=1.0.5
qdrant-client>=1.17.0
mcp>=1.26.0
pyyaml>=6.0
pydantic>=2.0
